Move Forms from application into module
